+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.1.0](https://github.com/sdsarun/no-more-random-ad/compare/v0.0.1...v0.1.0) (2025-05-27)
 ### [0.0.3](https://github.com/sdsarun/no-more-random-ad/compare/v0.0.2...v0.0.3) (2025-04-09)
 
 ### [0.0.2](https://github.com/sdsarun/no-more-random-ad/compare/v0.0.1...v0.0.2) (2025-04-08)
@@ -9,6 +10,10 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
+* add features random thai code ([0c974c2](https://github.com/sdsarun/no-more-random-ad/commit/0c974c23e8dbeb632ca2748ec92128ce9bb2530c))
+* **convert-thai-code:** done features ([2f97253](https://github.com/sdsarun/no-more-random-ad/commit/2f97253c6425a0d2f0669fc429435956b782ba5c))
+* **copy-history:** add pass props to config root container ([b82ef6c](https://github.com/sdsarun/no-more-random-ad/commit/b82ef6c732c20e0bbe4e287557cd152ddca45e3b))
+* **copy-history:** add use writeText first then fallback to original ([61da66c](https://github.com/sdsarun/no-more-random-ad/commit/61da66c0eb174f5271e6bb26d5332f5b6aad2b53))
 * **copy-history:** add pass props to config root container ([b82ef6c](https://github.com/sdsarun/no-more-random-ad/commit/b82ef6c732c20e0bbe4e287557cd152ddca45e3b))
 * **libs:** add label component ([804f999](https://github.com/sdsarun/no-more-random-ad/commit/804f9992d0b95141573814d236900003165ca808))
 * **libs:** add label component ([71271cb](https://github.com/sdsarun/no-more-random-ad/commit/71271cb5e80f954bc6cf3806b4f4ae4d260699f7))
@@ -16,6 +21,13 @@ All notable changes to this project will be documented in this file. See [standa
 * **random-thai-national-id:** add pass props to config root container ([fd3c529](https://github.com/sdsarun/no-more-random-ad/commit/fd3c52929077d350a12c646719ff4f70e875287c))
 * **random-thai-national-id:** add random national thai id ([8e7b3cf](https://github.com/sdsarun/no-more-random-ad/commit/8e7b3cfd992ea6cafa1a45b37a0cc4412bf96898))
 * **random-uuid:** add pass props to config root container ([adc5e04](https://github.com/sdsarun/no-more-random-ad/commit/adc5e0477af432a3f8b1f76cf104005d855ac1ac))
+* **ui:** add button select group ([c408a5b](https://github.com/sdsarun/no-more-random-ad/commit/c408a5bce3817400592b11add95d78df2fa63e22))
+* **ui:** add checkbox ([cd5e24b](https://github.com/sdsarun/no-more-random-ad/commit/cd5e24bfea7a4c0aa646bf0c5fd9f7715922ee9d))
+* **ui:** add code component ([fc6cc53](https://github.com/sdsarun/no-more-random-ad/commit/fc6cc53d6ac5afa58b7e09aebcd41e280b8ebfa4))
+* **ui:** add form ([919461a](https://github.com/sdsarun/no-more-random-ad/commit/919461acbcdc78e70a71df36fdf7a7fc78d81ef7))
+* **ui:** add input ([7be4ca6](https://github.com/sdsarun/no-more-random-ad/commit/7be4ca67e11126ffdada83d08c234cf35eca5422))
+* **ui:** add label component ([a72f01d](https://github.com/sdsarun/no-more-random-ad/commit/a72f01d84c797ed9001f2c2902e6bb80a42bb865))
+* **ui:** add radio group ([1a16266](https://github.com/sdsarun/no-more-random-ad/commit/1a162669687d3f033496a507bf475036b5b4006d))
 * **ui:** add checkbox ([cd5e24b](https://github.com/sdsarun/no-more-random-ad/commit/cd5e24bfea7a4c0aa646bf0c5fd9f7715922ee9d))
 * **ui:** add code component ([fc6cc53](https://github.com/sdsarun/no-more-random-ad/commit/fc6cc53d6ac5afa58b7e09aebcd41e280b8ebfa4))
 * **ui:** add label component ([a72f01d](https://github.com/sdsarun/no-more-random-ad/commit/a72f01d84c797ed9001f2c2902e6bb80a42bb865))
@@ -23,6 +35,10 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Bug Fixes
 
+* **convert-thai-code:** add responsive ([8bdd577](https://github.com/sdsarun/no-more-random-ad/commit/8bdd577a40d722b2753ef2ec732128c9e5ae0bd8))
+* **copy-history:** add responsive content ([adf7779](https://github.com/sdsarun/no-more-random-ad/commit/adf7779ad6c39e8b29c0e6caa52ae0631315d9f4))
+* **copy-history:** add responsive content ([20be1e3](https://github.com/sdsarun/no-more-random-ad/commit/20be1e37b8c2c0e34b2d9cb0885517165f135abb))
+* **copy-history:** wrap content with blob for various case ([df4ddb2](https://github.com/sdsarun/no-more-random-ad/commit/df4ddb27e6325b95543c2c9f9df935d26845aebb))
 * **random-uuid:** use code component instead ([34bc0eb](https://github.com/sdsarun/no-more-random-ad/commit/34bc0ebe06eed3eb6a745b2f0278e41157b34bb0))
 * **ui:** remove footer and add pass props ([e44e308](https://github.com/sdsarun/no-more-random-ad/commit/e44e308d3952f81a71fb7423dcf850dabaccb875))